Removed the old EnvironmentVariables class, as these values got moved to the Calamari.Testing library
