+++ source/Calamari.Tests/Fixtures/Commands/CommandFromModuleTest.cs
@@ -5,7 +5,7 @@ using System;
 using System.IO;
 using Calamari.Common.Plumbing.FileSystem;
 using Calamari.Common.Plumbing.Variables;
-using Calamari.Integration.FileSystem;
+using Calamari.Testing;
 
 namespace Calamari.Tests.Fixtures.Commands
 {
