+++ source/Calamari.Tests/Fixtures/Integration/Packages/GitHubPackageDownloadFixture.cs
@@ -3,9 +3,8 @@ using System.IO;
 using System.Net;
 using Calamari.Common.Plumbing.FileSystem;
 using Calamari.Common.Plumbing.Variables;
-using Calamari.Integration.FileSystem;
 using Calamari.Integration.Packages.Download;
-using Calamari.Integration.Processes;
+using Calamari.Testing;
 using Calamari.Tests.Helpers;
 using NUnit.Framework;
 using Octopus.Versioning.Semver;
